Split operator tests into separate files by operator type
Move ADD, SUB, and MUL operator tests out of vm.rs (1457 -> 748 lines)
and end_to_end.rs (577 -> 108 lines) into dedicated per-operator test
files. Shared test helpers extracted into common/ modules.

VM crate:
- vm/tests/common/mod.rs: VmBuffers, single_function_container, assert_trap
- vm/tests/execute_add_i32.rs: 2 ADD overflow/wrapping tests
- vm/tests/execute_sub_i32.rs: 11 SUB tests (basic, overflow, edge cases)
- vm/tests/execute_mul_i32.rs: 13 MUL tests (basic, overflow, edge cases)

Codegen crate:
- codegen/tests/common/mod.rs: VmBuffers, parse_and_run
- codegen/tests/end_to_end_add.rs: 5 ADD end-to-end tests
- codegen/tests/end_to_end_sub.rs: 9 SUB end-to-end tests
- codegen/tests/end_to_end_mul.rs: 9 MUL end-to-end tests
